Move MembershipFilter to its own file
